split up python build-scripts
- use environment variables in maven pom.xml
- remove former MacOS shell scripts
