+++ build_all.py
@@ -1,28 +1,29 @@
-import build_platform as b
+import build as b
+import build_constants as c
 
-# b.execute_build(b.target_macos, b.arch_x86, b.build_all, node_enabled = True, cross_compile = True)
-b.execute_build(b.target_macos, b.arch_x64, b.build_all, node_enabled = True, cross_compile = True)
-# b.execute_build(b.target_macos, b.arch_x64, [b.build_j2v8_jni, b.build_j2v8_java], node_enabled = True, cross_compile = True)
-# b.execute_build(b.target_macos, b.arch_x64, [b.build_j2v8_java], node_enabled = True, cross_compile = True)
+# b.execute_build(c.target_macos, c.arch_x86, c.build_all, node_enabled = True, cross_compile = True)
+b.execute_build(c.target_macos, c.arch_x64, c.build_all, node_enabled = True, cross_compile = True)
+# b.execute_build(c.target_macos, c.arch_x64, [c.build_j2v8_jni, c.build_j2v8_java], node_enabled = True, cross_compile = True)
+# b.execute_build(c.target_macos, c.arch_x64, [c.build_j2v8_java], node_enabled = True, cross_compile = True)
 
-# b.execute_build(b.target_linux, b.arch_x64, b.build_all, True, True)
+# b.execute_build(c.target_linux, c.arch_x64, c.build_all, True, True)
 
 # build Node.js only
 # def build_njs(target, arch):
-#     b.execute_build(target, arch, [b.build_node_js], node_enabled = True, cross_compile = True)
+#     b.execute_build(target, arch, [c.build_node_js], node_enabled = True, cross_compile = True)
 
-# build_njs(b.target_android, b.arch_arm)
-# build_njs(b.target_android, b.arch_x86)
+# build_njs(c.target_android, c.arch_arm)
+# build_njs(c.target_android, c.arch_x86)
 
-# build_njs(b.target_linux, b.arch_x86)
-# build_njs(b.target_linux, b.arch_x64)
+# build_njs(c.target_linux, c.arch_x86)
+# build_njs(c.target_linux, c.arch_x64)
 
 # # needs reboot here to turn Hyper-V off if Host-OS is Windows
 
-# build_njs(b.target_macos, b.arch_x86)
-# build_njs(b.target_macos, b.arch_x64)
+# build_njs(c.target_macos, c.arch_x86)
+# build_njs(c.target_macos, c.arch_x64)
 
 # # needs reboot here to switch to Windows-Containers
 
-# build_njs(b.target_win32, b.arch_x86)
-# build_njs(b.target_win32, b.arch_x64)
+# build_njs(c.target_win32, c.arch_x86)
+# build_njs(c.target_win32, c.arch_x64)
