+++ build_constants.py
@@ -1,0 +1,21 @@
+# target platforms
+target_android = 'android'
+target_linux = 'linux'
+target_macos = 'macos'
+target_win32 = 'win32'
+
+# target architectures
+arch_x86 = 'x86'
+arch_x64 = 'x64'
+arch_arm = 'arm'
+
+# core build-steps
+build_node_js = 'nodejs'
+build_j2v8_jni = 'j2v8jni'
+build_j2v8_java = 'j2v8java'
+
+# build-step aliases / aggregate steps
+build_all = 'all'
+build_full = 'full'
+build_native = 'native'
+build_java = 'java'
